Rediseno de algunas tablas para que se pueda usar eloquent orm en las relaciones one to many y vista Calendario para actividades

+++ database/migrations/2021_09_09_192248_create_adm_municipios_table.php
@@ -16,7 +16,7 @@ class CreateAdmMunicipiosTable extends Migration
         Schema::create('adm_municipios', function (Blueprint $table) {
             $table->id();
             $table->integer('adm_departamento_id');
-            $table->integer('cod_red');
+            $table->unsignedBigInteger('cod_red');
             $table->string('nom_provincia');
             $table->string('nom_municipio');            
             $table->integer('cod_municipio');
